Make Roster, skill Select Pannel
Roster를 ScrollView를 이용해 만듦.
스킬 선택창 만듦.
Roster의 캐릭터를 편성하면, 스킬 선택창이 켜짐.
Roster에서 캐릭터를 드래그 할 때, 캐릭터 오브젝트가 그대로 따라오지 않고 ghost를 만들어서 움직이게 함.

+++ Assets/Scripts/DataScript/CSVReader.cs
@@ -138,8 +138,8 @@ public class CSVReader : MonoBehaviour
             {
                 skillID = cols[0],
                 skillName = cols[1],
-                random_min = int.Parse(cols[2]),
-                random_max = int.Parse(cols[3]),
+                random_min = float.Parse(cols[2]),
+                random_max = float.Parse(cols[3]),
                 multiplier = float.Parse(cols[4]),
                 useMP = int.Parse(cols[5]),
                 coolTime = int.Parse(cols[6]),
